file_slotter
In this project , I have created  python file called file_slotter. Basically, work of this project is to put the file in respective folders.
I have created to 5 folders ( Images, Medias, softwares, docs and others). the file which are present in your personal system will directly goes in that 5 folders respectively like images goes in images and mp4, mp3, etc goes to medias folder and if you want to add another folder you have edit this project in python IDE.